@@ -16,97 +16,97 @@
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@agm/core": "1.0.0-beta.5",
-    "@angular/animations": "7.0.0",
-    "@angular/common": "7.0.0",
-    "@angular/compiler": "7.0.0",
-    "@angular/core": "7.0.0",
-    "@angular/forms": "7.0.0",
-    "@angular/http": "7.0.0",
-    "@angular/platform-browser": "7.0.0",
-    "@angular/platform-browser-dynamic": "7.0.0",
-    "@angular/router": "7.0.0",
-    "@fortawesome/fontawesome-free": "5.5.0",
+    "@agm/core": "1.1.0",
+    "@angular/animations": "8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "7.2.15",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@fortawesome/fontawesome-free": "5.12.0",
     "@ngx-translate/core": "11.0.1",
     "@ngx-translate/http-loader": "4.0.0",
-    "@swimlane/ngx-datatable": "14.0.0",
-    "ag-grid-angular": "19.1.2",
-    "ag-grid-community": "19.1.3",
-    "angular-tree-component": "8.0.0",
+    "@swimlane/ngx-datatable": "16.0.2",
+    "ag-grid-angular": "22.1.1",
+    "ag-grid-community": "22.1.1",
+    "angular-tree-component": "8.5.2",
     "angular2-text-mask": "9.0.0",
-    "angular2-toaster": "7.0.0",
-    "bootstrap": "4.1.3",
-    "chart.js": "2.7.3",
+    "angular2-toaster": "8.0.0",
+    "bootstrap": "4.4.1",
+    "chart.js": "2.9.3",
     "classlist.js": "1.1.20150312",
-    "codemirror": "5.42.0",
-    "core-js": "2.5.4",
+    "codemirror": "5.50.0",
+    "core-js": "3.6.1",
     "easy-pie-chart": "2.1.7",
-    "enhanced-resolve": "3.3.0",
-    "fullcalendar": "3.9.0",
+    "enhanced-resolve": "4.1.1",
+    "fullcalendar": "3.10.1",
     "intl": "1.2.5",
     "jqcloud2": "2.0.3",
-    "jquery": "3.3.1",
+    "jquery": "3.4.1",
     "jquery-slimscroll": "1.3.8",
     "jquery-sparkline": "2.4.0",
     "jquery.browser": "0.1.0",
     "jquery.flot.tooltip": "github:krzysu/flot.tooltip",
-    "lodash": "4.17.11",
-    "modernizr": "3.6.0",
-    "moment": "2.22.2",
-    "ng2-charts": "1.6.0",
+    "lodash": "4.17.15",
+    "modernizr": "3.8.0",
+    "moment": "2.24.0",
+    "ng2-charts": "2.3.0",
     "ng2-dnd": "5.0.2",
-    "ng2-file-upload": "1.3.0",
+    "ng2-file-upload": "1.4.0",
     "ng2-img-cropper": "0.9.0",
-    "ng2-material-dropdown": "0.10.1",
+    "ng2-material-dropdown": "0.11.0",
     "ng2-table": "1.3.2",
     "ng2-validation": "4.2.0",
-    "ngx-bootstrap": "3.1.2",
-    "ngx-chips": "2.0.0-beta.0",
-    "ngx-color-picker": "7.0.2",
-    "ngx-infinite-scroll": "6.0.1",
-    "ngx-select-ex": "3.6.7",
-    "popper.js": "1.14.3",
-    "rxjs": "6.3.3",
-    "rxjs-compat": "6.1.0",
-    "screenfull": "3.3.3",
+    "ngx-bootstrap": "5.2.0",
+    "ngx-chips": "2.1.0",
+    "ngx-color-picker": "8.2.0",
+    "ngx-infinite-scroll": "8.0.1",
+    "ngx-select-ex": "3.7.4",
+    "popper.js": "1.16.0",
+    "rxjs": "6.5.4",
+    "rxjs-compat": "6.5.4",
+    "screenfull": "5.0.0",
     "simple-line-icons": "2.4.1",
-    "source-sans-pro": "2.40.0",
-    "spinkit": "1.2.5",
-    "summernote": "0.8.10",
+    "source-sans-pro": "3.6.0",
+    "spinkit": "2.0.1",
+    "summernote": "0.8.14",
     "sweetalert": "2.1.2",
-    "ts-helpers": "1.1.1",
+    "ts-helpers": "1.1.2",
     "weather-icons": "github:erikflowers/weather-icons",
-    "web-animations-js": "2.2.1",
-    "zone.js": "0.8.26",
+    "web-animations-js": "2.3.2",
+    "zone.js": "0.10.2",
     "flot": "themicon/flot",
     "ika.jvectormap": "themicon/ika.jvectormap",
     "jquery.flot.spline": "themicon/jquery.flot.spline"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "7.0.4",
-    "@angular/compiler-cli": "7.0.0",
-    "@angular/language-service": "7.0.0",
-    "@types/node": "10.12.3",
-    "@types/codemirror": "0.0.67",
-    "@types/jasmine": "2.8.8",
-    "@types/jasminewd2": "2.0.3",
-    "@types/lodash": "4.14.118",
-    "codelyzer": "4.5.0",
-    "jasmine-core": "2.99.1",
+    "@angular-devkit/build-angular": "~0.803.21",
+    "@angular/cli": "8.3.21",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/node": "13.1.1",
+    "@types/codemirror": "0.0.82",
+    "@types/jasmine": "3.5.0",
+    "@types/jasminewd2": "2.0.8",
+    "@types/lodash": "4.14.149",
+    "codelyzer": "5.2.1",
+    "jasmine-core": "3.5.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "3.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "2.0.1",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "karma": "4.4.1",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "2.1.1",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.5.1",
     "karma-read-json": "1.1.0",
     "loaders.css": "0.1.2",
-    "protractor": "5.4.0",
-    "ts-node": "7.0.0",
-    "tslint": "5.11.0",
-    "typescript": "3.1.1",
-    "webdriver-manager": "10.2.5"
+    "protractor": "5.4.2",
+    "ts-node": "8.5.4",
+    "tslint": "5.20.1",
+    "typescript": "3.5.3",
+    "webdriver-manager": "12.1.7"
   }
 }
